README-update v2

+++ src/components/Deleter.jsx
@@ -1,20 +1,29 @@
 import React from "react";
 import { deleteCommentById } from "../utils/api";
+import style from "./Deleter.module.css";
 
 class Deleter extends React.Component {
   state = {};
 
   removeComment = () => {
-    const { comment_id, deleteCard } = this.props;
-    deleteCard();
-    deleteCommentById(comment_id);
+    const { comment_id, deleteCard, isDeleted } = this.props;
+    if (!isDeleted) {
+      deleteCard();
+      deleteCommentById(comment_id);
+    }
   };
 
   render() {
     const { username, author } = this.props;
-    if (username === author) {
-      return <i className="fas fa-trash-alt" onClick={this.removeComment}></i>;
-    } else return <p></p>;
+    return (
+      <div className={style.main}>
+        {username === author ? (
+          <i className="fas fa-trash-alt" onClick={this.removeComment}></i>
+        ) : (
+          <p></p>
+        )}
+      </div>
+    );
   }
 }
